Rename roles, profile and users; add explicit exports
