+++ lib/core/exports.ts
@@ -1,0 +1,2 @@
+export * from "./DeviceManagerConfiguration";
+export * from "./DeviceManagerPlugin";
