+++ lib/modules/asset/exports.ts
@@ -1,0 +1,5 @@
+export * from "./types/AssetContent";
+export * from "./types/AssetEvents";
+export * from "./types/AssetApi";
+export * from "./roles/RoleAssetsAdmin";
+export * from "./roles/RoleAssetsReader";
