+++ lib/modules/model/exports.ts
@@ -1,0 +1,3 @@
+export * from "./types/ModelApi";
+export * from "./types/ModelContent";
+export * from "./types/ModelDefinition";
