+++ lib/modules/shared/exports.ts
@@ -1,0 +1,3 @@
+export * from "./types/DigitalTwinContent";
+export * from "./types/Metadata";
+export * from "./types/EmbeddedMeasure";
