+++ lib/modules/shared/index.ts
@@ -1,6 +1,9 @@
 export * from "./types/DigitalTwinContent";
 export * from "./types/Metadata";
 export * from "./types/EmbeddedMeasure";
+export * from "./types/KuzzleRole";
 export * from "./utils/lock";
 export * from "./utils/flattenObject";
 export * from "./utils/ask";
+export * from "./utils/snakeCase";
+export * from "./Module";
